// src/main/resources/angular/src/app/api/model/segmentConf.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface SegmentConf { 
    /**
     * Segment id
     */
    segmentId: number;
    /**
     * Segment name
     */
    segmentName: string;
    /**
     * Wkt
     */
    geometry?: string;
    scheduleId: number;
    configurationId?: number;
    deletedDate?: string;
    cityId: number;
    maxTime: number;
    places: number;
    maxTimeInfinity: boolean;
    description: string;
    zoneTypeId: number;
    segmentCode: string;
}